+++ src/Sphere.cpp
@@ -3,14 +3,14 @@
 
 // Default constructor
 Sphere::Sphere() {
-	radius = 1.0;
-	centreVertex = glm::vec3(0, 0, 0);
+	radius = 1.0f;
+	centreVertex = glm::vec3(0.0f, 0.0f, 0.0f);
 	color = ColorDBL();
 	material = Material::LAMBERTIAN;
 }
 
 // Value constructor
-Sphere::Sphere(double _radius, glm::vec3 _centreVertex, ColorDBL _color, Material _material) {
+Sphere::Sphere(float _radius, glm::vec3 _centreVertex, ColorDBL _color, Material _material) {
 	radius = _radius;
 	centreVertex = _centreVertex;
 	color = _color;
@@ -20,39 +20,39 @@ Sphere::Sphere(double _radius, glm::vec3 _centreVertex, ColorDBL _color, Materia
 // Calculate the intersection of a ray and the surface
 // Return true if hits the surface
 // Hits circle 0, 1 or 2 times. Want the first hit. Assume 1 hit = 0 hits
-bool Sphere::rayIntersection(Ray& ray, double& minDistance) const {
+bool Sphere::rayIntersection(Ray& ray, float& minDistance) const {
 	glm::vec3 rayStart = ray.getStartpoint();
 	glm::vec3 rayDirection = glm::normalize(ray.getDirection());
 
 	glm::vec3 x_i; // Intersection point
 
 	// Get parameters for quadratic equation
-	double c1 = glm::dot(rayDirection, rayDirection);
+	float c1 = glm::dot(rayDirection, rayDirection);
 	float c2 = 2.0f * glm::dot(rayDirection, rayStart - centreVertex);
-	double c3 = glm::dot(rayStart - centreVertex, rayStart - centreVertex) - radius * radius;
+	float c3 = glm::dot(rayStart - centreVertex, rayStart - centreVertex) - radius * radius;
 
 	// Arg is the value under square root
-	double arg = c2 * c2 - 4.0 * c1 * c3;
+	float arg = c2 * c2 - 4.0f * c1 * c3;
 
 	// Ray misses Sphere
-	if (arg < 0) return false;
+	if (arg < 0.0f) return false;
 
 	// One solution, touches the side
 	if (arg < COMPARE_ELLIPSE) {
-		double t = -c2 / (2.0 * c1);
-		x_i = rayStart + rayDirection * (float)t;
+		float t = -c2 / (2.0f * c1);
+		x_i = rayStart + rayDirection * t;
 	}
 
 	// Time parameter for ray intersection, two possible solutions
-	double t1 = (-c2 + glm::sqrt(arg)) / (2 * c1);
-	double t2 = (-c2 - glm::sqrt(arg)) / (2 * c1);
+	float t1 = (-c2 + glm::sqrt(arg)) / (2.0f * c1);
+	float t2 = (-c2 - glm::sqrt(arg)) / (2.0f * c1);
 
 	// Take the parameter of the non negative solution, if any
 	if (t2 > 0.0f) {
-		x_i = rayStart + rayDirection * (float)t2;
+		x_i = rayStart + rayDirection * t2;
 	}
 	else if (t1 > 0.0f) {
-		x_i = rayStart + rayDirection * (float)t1;
+		x_i = rayStart + rayDirection * t1;
 	}
 	else {
 		return false;
@@ -68,7 +68,6 @@ bool Sphere::rayIntersection(Ray& ray, double& minDistance) const {
 	}
 
 	return false;
-
 }
 
 // Get what material the Sphere is made of
